requirements.txt
flask
pytesseract
Pillow
agno
groq
e2b_code_interpreter
python-dotenv
gunicorn
